Rydder litt i opprettelse av gjennomføring
